Fixes after review: DefaultResponseEncoder moved to standalone engine
