+++ engine/standalone/app/src/main/scala/pl/touk/nussknacker/engine/standalone/http/StandaloneRequestHandler.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.server.{Directive1, Directives}
 import cats.data.EitherT
 import cats.instances.future._
 import io.circe.Json
-import pl.touk.nussknacker.engine.standalone.StandaloneProcessInterpreter
+import pl.touk.nussknacker.engine.standalone.{DefaultResponseEncoder, StandaloneProcessInterpreter}
 import pl.touk.nussknacker.engine.standalone.api.types.GenericResultType
 import pl.touk.nussknacker.engine.standalone.api.{StandaloneGetSource, StandalonePostSource}
